Returns a dataframe summary of all existing Mercedes-Benz car listings Summary mileage, price, and rarity information is shown.

Returns a dataframe summary of all existing Mercedes-Benz car listings Summary mileage, price, and rarity information is shown.

listing_summary(
  data,
  model = "all",
  model_col = "model",
  odometer_col = "odometer_mi",
  price_col = "price_USD"
)

Arguments

data

data.frame The input Mercedes-Benz dataframe

model

character A character vector that lists all Mercedes-Benz model types to be included in the final summary table.

model_col

character The name of the column in the input dataframe that represents the Mercedes-Benz model

odometer_col

character The name of the column in the input dataframe that represents the Mercedes-Benz mileage (odometer reading in miles)

price_col

character The name of the column in the input dataframe that represents the Mercedes-Benz sale price (price in USD)

Value

data.frame Summary dataframe of Mercedes-Benz models, including average and median mileage (mi) and price (USD), and information about rarity. A model will be considered common, uncommon, or rare based on the percentage of market listings.

Examples

toy_data <- data.frame(model = c("gl-class", "gl-class", "gl-class", "amg", "amg", "amg"),
odometer_mi = c(100000, 200000, 400000, 100000, 200000, 400000),
price_USD = c(10000, 20000, 40000, 10000, 20000, 40000))
listing_summary(data = toy_data, model="all")
#> # A tibble: 2 × 7
#>   model    avg_mileage_mi median_mileage_mi avg_price_USD media…¹ perce…² rarity
#>   <chr>             <dbl>             <dbl>         <dbl>   <dbl>   <dbl> <chr> 
#> 1 amg              150000            150000         15000   15000      50 Common
#> 2 gl-class         150000            150000         15000   15000      50 Common
#> # … with abbreviated variable names ¹​median_price_USD, ²​percent_of_market
listing_summary(data = toy_data, model="gl-class")
#> # A tibble: 1 × 7
#>   model    avg_mileage_mi median_mileage_mi avg_price_USD media…¹ perce…² rarity
#>   <chr>             <dbl>             <dbl>         <dbl>   <dbl>   <dbl> <chr> 
#> 1 gl-class         150000            150000         15000   15000      50 Common
#> # … with abbreviated variable names ¹​median_price_USD, ²​percent_of_market